requirements.txt
customtkinter==5.2.2
Pillow==12.1.1
pyFirmata2==2.5.1
pyserial==3.5
bcrypt==5.0.0
SpeechRecognition 
#openai-whisper 
setuptools 
soundfile
pyaudio
transformers 
accelerate
noisereduce 
scipy 
numpy
faster-whisper
nvidia-cublas-cu12 
nvidia-cudnn-cu12
arabic-reshaper
python-bidi